refactor: Split ContentView monolith into per-view files
Reduce ContentView.swift from 2774 lines to the 109-line root view and
move the other 20 top-level types into 8 cohesive files: ModelStatusBar,
ChatViews, StatsViews, GenerateView, VisionView, FluxToolsView,
ModelsManagementView, SettingsView. Pure code movement, no behavior change.
